requirements.txt
asgiref==3.7.2
autopep8==2.0.4
Django==5.0.3
django-auditlog==2.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.1
django-quill-editor==0.1.40
djangorestframework==3.14.0
drf-nested-routers==0.93.5
Markdown==3.5.2
pillow==10.2.0
pycodestyle==2.11.1
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
whitenoise==6.6.0
